Add "javadoc" target to stroke build.xml file
This change adds a javadoc target to the build file for stroke, so
that you can do

% cd build/stroke
% ant javadoc

I have not updated any of the other build infrastructure (Makefile
etc.).  Nor have I added any intelligence to the "javadoc" task so
that it won't bother creating documentation if the doc is already up
to date etc..

Note that I've not used the "linkoffline" directive, which means that
if you try to generate doc when the Oracle web pages are offline,
things may hang until the connection times out.

Test-information:

"ant javadoc" works as expected
"isodebuild clean -component=stroke" gets rid of any javadoc