Add toString to some more classes
Also made "Stanza" be an abstract class and had its ".toString()" include
the name of the subclass which is involved, so that the subclasses don't
have to do that themselves.

Also added null check to existing HostAddress.toString() method

Also fixed Remko copyright in Connector class

Test-information:

Stuff is displayed as expected in debugger.