Update ByteArray with "readFromFile" implementation and more javadoc
This change provides implementation for the previously empty
"readFromFile" method, which is based on the code inside Swiften's
ByteArray.readByteArrayFromFile().

Note that, as for the Swiften method, no errors are reported if an
attempt to read from a file fails; it just leaves the contents of the
object unchanged.

Some javadoc was also provided, especially noting that the various
String methods are only safe so long as you're working with UTF-8
Strings.

Test-information:

Existing Stroke unit tests still pass

Signed-off-by: Nick Hudson <nick.hudson@isode.com>