Implement "CertificateWithKey" and add support for setting client certificates
This change provides the functionality to allow clients to specify a
PKCS#12 file containing client certificate/key for use when starting
TLS sessions.

The PKCS12Certificate class now subclasses "CertificateWithKey"
(matching the Swiften implementation).

Swiften also has "CAPICertificate", which is another subclass of
CertificateWithKey. This has not been provided in this patch.

From a client's point of view, all that's necessary to specify a
certificate to be used for TLS is to do something like

  CertificateWithKey myCert = new PKCS12Certificate(
           "/home/fred/myp12file.p12",
           "secret".toCharArray());

  coreClient.setCertificate(myCert);

before calling "CoreClient.connect".

Matching the Swiften functionality, constructing a new
PKCS12Certificate does not actually perform validation of the P12
file/passphrase; that takes place when the p12 file is used.

There is limited scope for returning to the caller errors describing
possible problems, but JSSEContext uses the "emitError" method which
does maintain error information, which is available in a debugger, or
from the JSSEContext.toString() method.

Test-information:

Set up an M-Link server with TLS verified that

 - when I specify a client certificate with suitable SAN, the client
   sends it and the server reports authentication using the certificate

 - when I specify a client certificate without a suitable SAN, the
   client sends it but the server rejects it