Port Error Parser to stroke
After this change, the error payload object should be populated in case of error.
The condtion, type and text field will be from the payload rather than Undefined,
Cancel and empty.

Test-information:

tested by executing adhoc-commands on an XMPP clinet in a way to result
in an error. I do see the error text and condition set as per the XMPP streams.

Reviewer: Kevin Smith <kevin.smith@isode.com>