Remove "max" limit on buffer size for SSL connections
The original implementation of JSSEContext allocated buffers for the
SSL data which started off at a size determined using information from
the SSLSession, and which were allowed to grow to up to ten times
their original size.

When testing with jabber.org it was fairly easy to exceed this size
(e.g. requiring a buffer of ~650K where the maximum value had been
around 160K), meaning that applications would fail.

This change removes the upper limit altogether.  Now, the buffer will
grow to whatever is required, so long as free memory is available.

I also renamed "enlargeBuffer" in response to a previous review comment

Test-information:

No longer see problems when talking to jabber.org over
SSL. Instrumentation shows that buffer is growing as expected.