Fix PlatformDomainNameResolve DomainNameAddressQuery
In the PlatformDomainNameResolver class there is a DomainNameAddressQuery
class (accesible via DomainNameResolver->createAddressQuery()) for
performing a DNS lookup on a given domainname.  This should have been
returing the set of all HostAddress associated with a given domain, but
instead was only returning a singleton set (or empty if there was no dns).

This patch fixes this by changing the method call from
InetAddress.getByName() to InetAddress.getAllByName().

Test-information:

Tested on top of my MLC Diagnose SRV patch.  For 'google.com' we now see a
full list of ip addresses associated with it, rather then just the one.

Change-Id: I6e57c16bb64f76048f16bcff8ee9c1924049a051