Enable building of jar file without examples
The example code includes references to Swing, which isn't available
for all environments (e.g. Android) and so this change provides an
alternate build target to allow stroke.jar to be built without
processing the example code.

The original "dist" target was incorrect in the way it was creating
the jar file, because it was creating a standalone MANIFEST.MF file
(which didn't get used for the jar file at all).  So that has been
corrected (for the dist-with-examples target).

So if now do
   % ant -Dnoexamples=1
Tnen no examples will be built.

If you do
   % ant

then the jar file will include examples (as before) and will also have
a manifest that specifies "Main-class" properly.

Note that this change has already been made to the isode repository
and will not need applying there.

Test-information:

Tested building with/without examples. The jar file appears correct in
each case.

Prior to this patch, saying "java -jar stroke.jar" did not work,
because the manifest wasn't being used properly. After this patch, it
does (for the jar file that includes examples).

Change-Id: I68eadc4355cb655dd31e6afec48405a6fe2c057e
Signed-off-by: Nick Hudson <nick.hudson@isode.com>