Move hardcoded XMPP SRV information from Connector into CoreClient
The Connector class had "_xmpp-client._tcp." hard-coded in it, which
meant that it was not suitable for non XMPP clients.

This change means that Connector could now be used by clients who are
interested in arbitrary SRV records; the CoreClient class is updated
accordingly.

Test-information:

Built and tested using MLC.

Also tested with a client that is interested in IMAP SRV records

Change-Id: Ia23c148fd8afdd7b3271c47b1c96d086d57a44bd