Fix incorrect initialisation of jid_ field inside IQRouter
The class IQRouter has a private "jid_" field that was not being
initialised to contain an invalid JID, which meant that there was a
risk of NullPointerException if anyone called the "getJID()" method
and tried to use the returned JID.

This showed up because one of the unit tests was getting a
NullPointerException, which caused the failure:

  [junit] Test com.isode.stroke.queries.requests.GetPrivateStorageRequestTest FAILED

The failure was shown to have been introduced by the change "Check
sender on incoming IQ responses"
(535e1a979a164f807aa64bf2df2bb36e7015ff17)

This change fixes the initialisation.  The other fields in this class
are always initialised so can never be null.

Test-information:

After this patch, unit tests no longer show the failure.

Change-Id: Idfcabf5393c8353194dddc414d58c37301487908