Fix JavaConnection to emit onDataWritten when it writes data to the socket
The java code was never emitting the onDataWritten signal, although
the corresponding C++ code in Swiften does do this.

This change causes the signal to be emitted whenever a data is
successfully written to the socket.

Test-information:

Tested using an application which was registering for the signal;
previously it never saw "onDataWritten"; now it does.

Tested using an application which doesn't register for the signal; it
works as before.

Change-Id: I1399af0721ef8226c0c4d2420bbe23f53ad3494f