Don't use SSLv3 in JSSEContext
The POODLE vulnerability means that using SSLv3 is insecure. So this
change removes it from the list of protocols that JSSEContext may use.

Oracle's "Java Cryptography Architecture Standard Algorithm
Name Documentation"

  http://docs.oracle.com/javase/7/docs/technotes/guides/security/StandardNames.html

Lists the "standard names" that can be used in this context:

 SSLv2
 SSLv3
 TLSv1
 TLSv1.1
 TLSv1.2
 SSLv2Hello

After this patch, only the three "TLS" protocols will be allowed.

Test-information:

Tested using JRE6 and JRE7; viewing the SSL handshake indicates that
the protocol being requested is being used when the handshake occurs

Change-Id: I99710a72a4b8567226b1205fdf64c6c67ccc2a9a