Avoid potential double-adding of "lastConsumed" count to "bytesConsumed"
The code is currently doing
  bytesConsumed += lastConsumed;

inside one of the clauses in a while loop, and then the same operation
again when the while loop exits.  So there is a risk that the
"bytesConsumed" value will be too large.

In fact the only place the value of "bytesConsumed" is used is by some
code that checks whether it's greater than zero, so it would not be
problematic if the value were too large.

It seems worth fixing this in case future changes rely on the
"bytesConsumed" value being accurate.

Test-information:

inspection only

Change-Id: Ibd6fd01417afc4c4e030a5173bfba9a02980a757
signed-off-by:robert.williams@isode.com