Make sure Stroke clears write buffer before disconnecting
There was a bug with stroke where if I sent a large amount of messgaes (cica
1000) and then disconnected, not all the messages would be sent before storke
handled the disconnect.  It seems it did not fully lear the write buffer before
closing the stream.

This patch shoudl fix the error. It now only closses if the disconnecting flag
is closed and the write buffer is closed.  This should bring it into line with
Swiften behaviour.

Test-information:

I had a simple stroke application that sent 1000 messages and then disconnected.
Before the patch it would only send abut 80 or so of the messages before it got
disconnected.  After the patch it now sends all the messages before
disconnecting.

Change-Id: I6f56b25dcbabd16ee860dbf353f39559a77d6830
Signed-off-by: Alex Clayton <alex.clayton@isode.com>