Add StringIndexOutOfBounds catch to IDNA
When run with java 7  class IDNA.getEncoded was throwing a StringIndexOutOfBounds exception if it was passed a String begining with '.isode'.  This has the knock on effect of some of one of the isode MLC tests failing.   This was caused by IDN.toAscii throwing the exception, when it should only thrown IllegalArgumentException.

Java 7 is no longer supported so I can't raise a bug for this, a fix for stroke is easy though, just add the StringIndexOutOfBoundException to the catch block.

Test-information:

Ran the tests that were failing in eclipse on my machine (which runs java 7) tests now all passed correctly.

Sanity test that tests still passed when run on command line against java 8.

Ran the Stroke JUnit test against the code.  Tests all passed when using latest version of java 7 (7_79) and when using java 8.

Change-Id: Ifc9c830be7a0e5c9e5f2330a6782eff2401f18cb