Fix ClientSession class bug due to recent stroke changes
Stroke was brought in line with swiften as a result of which strong
(certificate based bind) authentication between client and xmpp server
stopped working. This was happening because of missing "=" in the
<auth xmlns='urn:ietf:params:xml:ns:xmpp-sasl' mechanism='EXTERNAL'>=</auth>

The AuthRequestSerialiser was not adding "=" because the ByteArray object
that was passed was null.

This patch fixes the ClientSession class to pass empty ByteArray object as
is done in existing swiften code.

Test-information:

strong bind works using java client to xmpp server

Change-Id: Ide2d61723d92df19a251db7f1eb9f1581329f782